Relocate RemoteButler registry-only functions
To remove some unnecessary circular dependency, move private functions used only by RemoteButlerRegistry off of RemoteButler into RemoteButlerRegistry.
